requirements.txt
annotated-types==0.7.0
anyio==4.4.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
fastapi==0.115.0
google-api-core==2.19.2
google-api-python-client==2.146.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.65.0
h11==0.14.0
httplib2==0.22.0
idna==3.10
oauthlib==3.2.2
proto-plus==1.24.0
protobuf==5.28.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
pyparsing==3.1.4
python-dotenv==1.0.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
sniffio==1.3.1
starlette==0.38.5
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.30.6
openai==1.56.1
httpx==0.27.2
paddlepaddle
paddleocr
fastapi[all]
python-jose[cryptography]
passlib
python-dotenv
bcrypt==3.2.0

